xmr_mining_all_options_with_test
